remove console
