+++ backend/src/weca_client/utils/logger.py
@@ -1,11 +1,9 @@
 import logging
-from rich.console import Console
 from rich.logging import RichHandler
 from logging import StreamHandler
 LOGGER_LEVEL = "DEBUG"
 LOGGER_MOD = "local"
 
-console = Console()
 
 
 if LOGGER_MOD == "local":
@@ -25,9 +23,3 @@ else:
     )
 
 log = logging.getLogger("csv_handler")
-# Usage:
-# log.debug({"key": "value", "key2": "value2"})
-# log.info("this is info message")
-# log.warning("this is warning message")
-# log.error("this is error message")
-# log.critical("this is critical message")
